+++ src/read.cpp
@@ -10,7 +10,7 @@
 
 namespace RALA {
 
-std::unique_ptr<Read> createRead(uint32_t id, const char* name, uint32_t name_length,
+std::unique_ptr<Read> createRead(uint64_t id, const char* name, uint32_t name_length,
     const char* sequence, uint32_t sequence_length, const char* quality,
     uint32_t quality_length) {
 
@@ -21,7 +21,7 @@ std::unique_ptr<Read> createRead(uint32_t id, const char* name, uint32_t name_le
         sequence_length, quality, quality_length));
 }
 
-Read::Read(uint32_t id, const char* name, uint32_t name_length, const char* sequence,
+Read::Read(uint64_t id, const char* name, uint32_t name_length, const char* sequence,
     uint32_t sequence_length, const char* quality, uint32_t quality_length)
         : id_(id), name_(name, name_length), sequence_(sequence, sequence_length),
         quality_(quality, quality_length), rc_() {
